requirements.txt
asgiref==3.8.1
Django==3.2.7
django-environ==0.6.0
django-import-export==2.5.0
djangorestframework==3.12.4
psycopg2==2.9.1
raven==6.10.0
sqlparse==0.4.4
typing_extensions==4.10.0
